Change attitude estimation algorithm

+++ Firmware/App/Src/bmi088.c
@@ -99,7 +99,7 @@ void bmi088_read_acc(float acc[3]) {
 
     for (int16_t i = 0; i < 3; i++) {
         x = pack2(data[2 * i + 1], data[2 * i]);
-        acc[i] = to_int16(x) / 32768.f * acc_range_sf[bmi088_config.acc_range] * ACC_GRAV;
+        acc[i] = (float)to_int16(x) / 32768.f * acc_range_sf[bmi088_config.acc_range] * ACC_GRAV;
     }
 }
 
@@ -111,7 +111,7 @@ void bmi088_read_gyro(float ang[3]) {
 
     for (int16_t i = 0; i < 3; i++) {
         x = pack2(data[2 * i + 1], data[2 * i]);
-        ang[i] = to_int16(x) / 32767.f * gyro_range_sf[bmi088_config.gyro_range] * DEG_TO_RAD;
+        ang[i] = (float)to_int16(x) / 32767.f * gyro_range_sf[bmi088_config.gyro_range] * DEG_TO_RAD;
     }
 }
