+++ Firmware/App/Src/lis2mdl.c
@@ -45,6 +45,6 @@ void lis2mdl_read_mag(float *mag) {
     /* Read magnetic field. */
     for (int16_t i = 0; i < 3; i++) {
         x = pack2(data[2 * i + 1], data[2 * i]);
-        mag[i] = to_int16(x) * SENSITIVITY;
+        mag[i] = (float)to_int16(x) * SENSITIVITY;
     }
 }
